# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "NFS API Server",
        "title": "NFS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "nfs-api.andasy.dev",
    "basePath": "/",
    "paths": {
...
        "/api/v1/invoices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "List all invoices",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter from date (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to date (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by user ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by service IDs (comma-separated)",
                        "name": "service_ids",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by notary ID",
                        "name": "notary_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by user status (paid/unpaid)",
                        "name": "user_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by notary status (pending/signed_unpaid/signed_paid)",
                        "name": "notary_status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/utils.PaginatedResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Create a new invoice",
                "parameters": [
                    {
                        "description": "Create payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/invoice.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/invoice.Invoice"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/invoices/user-summaries": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Get invoice summaries grouped by user",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter from date (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to date (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by notary ID",
                        "name": "notary_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by notary status",
                        "name": "notary_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by user status (paid/unpaid)",
                        "name": "user_status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/invoice.UserSummary"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/invoices/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Get invoice by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/invoice.Invoice"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Update an invoice",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/invoice.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/invoice.Invoice"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Delete an invoice",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/invoices/{id}/services/{serviceId}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Update invoice service payment status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Invoice Service ID",
                        "name": "serviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/invoice.UpdateServiceStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/invoice.Invoice"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "invoice.CreateRequest": {
            "type": "object",
            "required": [
                "client_name",
                "date",
                "notary_id",
                "services",
                "upi"
            ],
            "properties": {
                "client_name": {
                    "type": "string",
                    "maxLength": 150
                },
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "notary_id": {
                    "type": "integer"
                },
                "notary_status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "signed_unpaid",
                        "signed_paid"
                    ]
                },
                "note": {
                    "type": "string"
                },
                "ref_no": {
                    "type": "string",
                    "maxLength": 50
                },
                "services": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/invoice.CreateServiceItem"
                    }
                },
                "upi": {
                    "type": "string",
                    "maxLength": 40
                }
            }
        },
        "invoice.CreateServiceItem": {
            "type": "object",
            "required": [
                "copy",
                "name",
                "office_service_id",
                "price"
            ],
            "properties": {
                "copy": {
                    "type": "integer",
                    "minimum": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 100
                },
                "office_service_id": {
                    "type": "integer"
                },
                "payment_status": {
                    "type": "string",
                    "enum": [
                        "paid",
                        "unpaid"
                    ]
                },
                "price": {
                    "type": "number",
                    "minimum": 0
                }
            }
        },
        "invoice.Invoice": {
            "type": "object",
            "properties": {
                "client_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "notary": {
                    "$ref": "#/definitions/notary.Notary"
                },
                "notary_id": {
                    "type": "integer"
                },
                "notary_status": {
                    "description": "pending | signed_unpaid | signed_paid",
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "office": {
                    "$ref": "#/definitions/office.Office"
                },
                "office_id": {
                    "type": "integer"
                },
                "paid_amount": {
                    "type": "number"
                },
                "ref_no": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/invoice.InvoiceService"
                    }
                },
                "total_amount": {
                    "description": "Computed at runtime",
                    "type": "number"
                },
                "unpaid_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "upi": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/auth.User"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_status": {
                    "type": "string"
                }
            }
        },
...
        "invoice.UpdateRequest": {
            "type": "object",
            "properties": {
                "client_name": {
                    "type": "string",
                    "maxLength": 150
                },
                "date": {
                    "type": "string"
                },
                "notary_id": {
                    "type": "integer"
                },
                "notary_status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "unpaid",
                        "paid"
                    ]
                },
                "note": {
                    "type": "string"
                },
                "ref_no": {
                    "type": "string",
                    "maxLength": 50
                },
                "upi": {
                    "type": "string",
                    "maxLength": 40
                }
            }
        },
        "invoice.UpdateServiceStatusRequest": {
            "type": "object",
            "required": [
                "payment_status"
            ],
            "properties": {
                "payment_status": {
                    "type": "string",
                    "enum": [
                        "paid",
                        "unpaid"
                    ]
                }
            }
        },
        "invoice.UserServiceSummary": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "office_service_id": {
                    "type": "integer"
                },
                "paid_amount": {
                    "type": "number"
                },
                "total_amount": {
                    "type": "number"
                },
                "unpaid_amount": {
                    "type": "number"
                }
            }
        },
        "invoice.UserSummary": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "paid_amount": {
                    "type": "number"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/invoice.UserServiceSummary"
                    }
                },
                "total_amount": {
                    "type": "number"
                },
                "unpaid_amount": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "notary.Notary": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted": {
                    "description": "0 = active, 1 = deleted",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "names": {
                    "type": "string"
                },
                "office": {
                    "$ref": "#/definitions/office.Office"
                },
                "office_id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "description": "active | inactive",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.PaginatedResponse": {
            "type": "object",
            "properties": {
                "items": {},
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}